test(routing): consolidate and streamline fallback tests
